Merged PR 214424: Fix for backplane service thread pool overflow during service startup
What is happening today is that the backplane service get bombarded by thousand of request to update contacts during the startup where most of our clients are reconnecting causing the creation of thousand of threads to be 'forget' and later side effects on the thread pool.
We need a mechanism to 'buffer' and parallel those task in a controlled way. This PR will use the data flow library to achieve this purpose.

Also take the opportunity to disable the Cosmos until we fix the throughput on the contacts collection.