Merged PR 210228: fix: to make environment names case insensitive.
This is to prevent the user to create environments with different cases. However, they are allowed to create with lower/upper case for the first time.

earlier we were able to create an environment as 'a' and 'A'.
With this fix, we can now validate the environment names irrespective of the case at the client-side.