Merged PR 216601: add tags to image and image version
Added tags to image and image version so we can tell where VS and Live Share came from for those images. The original tags are created by the nexus dogfood tool now on the VM it creates, and the tags are propagated to the image by the inline script in the [vsclk-core-new-image](https://dev.azure.com/devdiv/OnlineServices/_releaseDefinition?definitionId=83&_a=definition-tasks&environmentId=313) pipeline. Then this script propagates those same tags to the images and image versions it creates.

Here's an example of the tags for today's images:

| Name | Value |
| ---- | ---- |
| LSBuildUrl | https://dev.azure.com/devdiv/DevDiv/_build/results?buildId=3272770 |
| VSBuildUrl | https://dev.azure.com/devdiv/DevDiv/_build/results?buildId=3272770 |
| VSSetupUrl | https://vsdrop.corp.microsoft.com/file/v1/Products/DevDiv/VS/9f5729807197d7104cef97ee036f47d6ee2bb5dc/1be8fb9d-d3cd-107e-0324-a295f747bafe%3bbootstrappers/Enterprise/vs_enterprise.exe |

> Note that LSBuildUrl and VSBuildUrl are the same now because we're not installing Live Share separately right now--we're using the one that comes with VS (from rel/d16.5).

Unrelated to the tags, I removed the work around for the problem mentioned in https://developercommunity.visualstudio.com/content/problem/752118/az-cli-reports-warnings-that-breaks-releases.html because the latest response indicates it's a problem with the version of powershell installed in the hosted 2019 agents. Following the recommendation there I switched to hosted 2016 agents and the problem is gone.

Related work items: #1025473