Merged PR 217929: Making fewer calls to the DB and using less RUs
Two changes to reduce the collection RU cost.
The older algorithm used to do two queries. One to get the last Summary and one to get all billing events within a certain time range. Now I've combined those queries
Now we get all billing events for the last 48 hours for our plan. We'll filter the returned items to get the last summary and then do a later filter (on the full set) to get all the environment state changes that happened in the last billable time period.

Secondly, I made the exact query we fire off to be more explicit. In my experimentation, this had a lower RU cost.