Merged PR 222250: [Front End] Fix Uris for environment creation, add tests
In my refactor of the EnvironmentController I made a helper to get the service and callback Uris for environment creation and start.  The helper didn't properly handle Urls for create which didn't end in a `'/'` (i.e. `/api/v1/environments` is busted, but `/api/v1/environments/` was OK).

Fixed the logic in the helper and added unit tests to catch this going forward.