Merged PR 222430: Fix auto-shutdown from heartbeats
Fix auto-shutdown from heartbeats

The second record update in the DB was failing with a precondition failure since the object being sent for the update was the original one. Ensure that the first update is fully captured so that the shutdown update is applied in order.

A better longer term fix will come in a separate PR, which is to have the manager wrap all DB operations in retries and to refetch and reapply object mutations if an update fails due to precondition failures.