Merged PR 221580: [FrontEnd] Fix telemetry for environment updates
The telemetry for the target SKU and AutoShutdown settings is currently being eaten by the logging for getting the available settings updates (`GetEnvironmentAvailableSettingsUpdates` is used by `UpdateEnvironmentSettingsAsync`  for validation and that method does logging which causes the temporary values to get dropped).

Reworked the `UpdateEnvironmentSettingsAsync` method a little to use `logger.OperationScopeAsync` so that the target settings telemetry will get logged correctly using the child logger.

Telemetry from the method will look like this now.  Note that for both success and failure the `msg` is the same, the difference is the existence of the `ErrorDetail` field.

- Successful update:
```
{
    "msg": "cloudenvironmentmanager_update_environment_settings_async_complete",
    "AutoShutdownDelay": "30",
    "SkuName": "standardLinux",
    "TargetAutoShutdownDelay": "120"
}
```

- Failed update:
```
{
    "msg": "cloudenvironmentmanager_update_environment_settings_async_complete",
    "AutoShutdownDelay": "30",
    "SkuName": "standardLinux",
    "TargetAutoShutdownDelay": "12031098",
    "ErrorDetail": "Error MessageCodes: [ RequestedAutoShutdownDelayMinutesIsInvalid ]"
}
```