Merged PR 222952: Support .NET core 3.1 for the signalR client nuget package
SInce we move our services to 3.1 we need a newer nuget package to be able to connect from a .NET 3.1.x client.
This is blocking to consume the latest version we produce from the vsclk-core repo to liveshare.

I confirm the CLI sample is able to connect to the new services.