Merged PR 224637: Fix json serialization on the signalR service for presence and the contact backplane
Today production deployment is broken for several entry points on the signalR presence service.
The default serialization of the new 3.1 will use the new JSON objects when we pass 'complex' objects like dictionaries that are very common on our presence API.
The solution was to force the use of the Newtonsoft library and so all our complex parameters will pass correctly into our existing and new clients.

The CLI was updated to catch the problem by sending a complex object on the message payload instead of passing a simple string, this will help to diagnosis similar problems in the future.