Merged PR 224227: Adding cache control header for 2 hours
Adding cache control header with MaxAge 2 hours for all the requests other than the ones starting with "/static".
The reason for adding this check is to differentiate between Document and JavaScript requests that are served by the service. Did not find a better way to differentiate.

Found that this is being returned for these requests as well:
- All index/document calls
- service-worker.js
- /configuration api-call
- auth callback with /?code=&client_info=&state=&session_state=
- /authenticate-port-forwarder

![image.png](https://devdiv.visualstudio.com/579bc0c8-404b-48fd-a86e-9d65b001f1bf/_apis/git/repositories/b262268f-bbd7-4d01-b7dc-b03666e77189/pullRequests/224227/attachments/image.png)