Merged PR 224774: [msal fix cookie issues]: remove msal cookie tracking
MSAL has an option to track auth request state in browser cookies, it was implemented as a fix for authentication loops in old IE and Edge browsers.

The problem with the cookie is that MSAL [sets it](https://github.com/AzureAD/microsoft-authentication-library-for-js/blob/dev/lib/msal-core/src/cache/BrowserStorage.ts#L37) but never [cleans it up](https://github.com/AzureAD/microsoft-authentication-library-for-js/blob/dev/lib/msal-core/src/cache/BrowserStorage.ts#L59). This creates more and more cookies in the user browser which eventually leads to few errors:

1. "The page is not reachable" error we see sometimes.
2. When there are too much cookies (and since they added as a header to every HTTP request) any API http request fails due to the header length limits (limits vary across browsers and web servers 4-18Kb).

This PR removes the option since we don't support the legacy browsers anyways. Also cleans up the related cookies to on initialization to recover them.