Merged PR 225071: Fix CORS bootstrap to allow VSO online origins to be allowed for liveshare web extension
On todays deployment the liveshare web extension refuse to connect to our signalR extension. The reason is that the 'preflight' request will eventually fail since the browser will pass the origin header and the signalR hub negotiate will fail due to the CORS security.
This PR will enable CORS support and will allow the vs online URIs to be allowed on any method and header.

The CORS origin is being defined on our app settings json to facilitate future changes.