Merged PR 226102: Do not cleanup storage during suspend.
Someone has added code to cleanup storage resource during suspend. This has caused the suspend operation to always fail, as the backend doesn't supporting cleaning up storage. So any suspend command to the environment, was resulting in error and leaving the database state to be in ShuttingDown state. This was causing a lot of bugs on client state changes. An environment entered clean shutdown state, only with a forcesuspend from the user or during server idle detection.

Removing the storage cleanup for suspend in this PR. This can also be fixed in ```CleanupResourceContinuationHandler``` to support storage resource cleanup, and leave the frontend call here. But that's inefficient resulting in extra unnecessary http requests.