Merged PR 224973: fix: telemetry event changes to find create environment times in VSO
As of today, we are missing provisioning time in the VSO client telemetry.

In order to form a kusto query to find the environment creation time with the state update event so that we can compute the creation duration in VSO.
we have to pass the environment id from 'creationSuccess' event. This has environment ID which we can use to tie up with stateChange event (which has environment Id)
We also want to add old state and new state to 'StateChange' event. Since we would like to know the state changes of an environment as well.

This will help us to come up with kusto query to find the actual creation time for an environment in VSO.