Merged PR 226750: Update image versions dynamically in the database
Add support for updating image versions dynamically in the database.

## Overview
* Change `SkuCatalog` to dynamically resolve VM agent blob urls and VM image versions dynamically
* Update the scaling engine to re-evaluate the SkuCatalog every minute and then recompute the set of active pools and their versions

## Design
Since the `SkuCatalog` is shared between frontend and backend, and since some of the version information ends up surfaced in the frontend, we need to keep the evaluated catalog (eventually) consistent between frontend and backend. But frontend and backend have different databases, so we can't use `SystemConfiguration` out of the box for dynamic overrides, otherwise they would each look for overrides in separate places. In order to resolve this, I've introduced an `ICurrentImageInfoProvider` interface which is used to resolve image names and versions dynamically. This allows us to have different implementations for frontend and backend.

To keep frontend and backend consistent, and to allow image updates to apply to a specific control plane region instead of globally, the backend implementation of `ICurrentImageInfoProvider` looks for overrides in the backend database using `SystemConfiguration`. The backend also exposes an API of this implementation for the frontend to call. The frontend has a delegated and cached implementation that calls into the backend to query the current image name/version properties.

_Frontend -> Backend -> Configuration DB_

If you apply an updated record in the DB, it should be visible to both the frontend and backend within one minute due to the caching intervals that are set up in both services.

## Database record format
In the **backend** control plane database, in the `configuration` collection, add a document with the following schema. Examples of each image type are shown below. Look at `appsettings.images.json` for ideas around what kind of image names and values you can override. The general `id` key format is `images:<familyType>:<familyName>:<property>` where `property` is either `name` or `version`.

### VM image version
```json
{
  "id": "images:compute:nexusInternalWindowsImage:version",
  "value": "2020.0203.XYZ"
}
```

### VM agent version (called "name" in our `appsettings.images.json`)
```json
{
  "id": "images:vmagent:vsoagentlinux:name",
  "value": "VSOAgent_linux_123456.zip"
}
```

### Kitchensink blob
```json
{
  "id": "images:storage:kitchenSink:name",
  "value": "cloudenvdata_kitchensink_1.0.1234-g<commit>.release123"
}
```

## Miscellaneous
* Unit tests all pass
* Manual testing of frontend and backend individually and together both seem to have everything working. Updates applied in the backend DB are surfaced in both places as expected.

Related work items: #1057288