Merged PR 228318: chore: excluding the static files for modernCop scan
excluding the static files for modernCop scan to perform playwright test.
We are getting 'moderncop' scan errors for newly added js files for playwright setup.
This is causing the builds to fail and not triggering the DEV deployments for VSO portal.

you can see the modercop scan issues over here
https://devdiv.visualstudio.com/OnlineServices/_build/results?buildId=3460108