Merged PR 228332: Refactor backplane services to fully support the relay hub
We made big investment in the past to support a backplane service that will host multiple signalR services on a kubernetes cluster so we can scale well our contacts presence service. The backplane service is capable to sync a signalR hub service with other instances running in the same cluster or in another region if needed. This was working perfectly fine except it only support the contact service hub. Since we also want to support the relay hub service we would need to offer the same replication capabilities to this service. This PR will refactor several pieces to achieve this goal. For example the json RPC server was refactored to support managing several target service on the backplane, instead of just the contacts. Also the redis connection multiplexer and the cosmos connection would need to share their entry point with multiple service providers,
Same for the signalR side, we need to offer a way to connect an rpc session on the same channel.

There are also several fixes on the client API for the relay proxy and also the equivalent typescript version.
This PR also enable the code analyzer and so many coding style changes were forced due to build errors.

A new 'relay' stress app was created in order to test perf when sending/receiving data trough the relay hub. The CLI will create 2 endpoints and transfer a MB file (configurable) using chunk data. The preliminary results are around 20 secs for a 1MB file to transfer data between one endpoint to another one.