Merged PR 228919: chore: Added PF Service base
This is a boilerplate setup for a service to be. Right now it doesn't really do much useful things.

I wanted to get the setup out of the way first so the reviews of the useful things can focus on that - the useful things.

I've added setup for VS SaaS SDK, Testing project with a few boilerplate tests so people can hop on and start writing tests.

This doesn't have ingress rules, or interact with existing services. Once merged, I'll add it to our deployment pipelines and we can start bringing in functionality in a verifiable manner.

Related work items: #1066699