Merged PR 230195: feat: ability to provide the user alias for devstamp
feat: ability to provide the user alias for devstamp

vsonline environments has vsonline as the user. by using the defaults, dogfooding developers would step on each others toes.
This allows specification of alias as a parameter in the appsettings.json file, and if vsonline alias is used it will throw.