Merged PR 232313: chore: to purge obsolete vso agent and kitchen sink images
**Description**
This pull request has code to have a worker that fires task for purging VSO agent and kitchen images.

We have been kept accumulating the images in the storage account in Azure.
So this worker is configured to run the task to delete the obsolete images that are a month old and also not being currently used in the respective regions.

**TODO:**
to add unit testing for the delete tasks.

Related work items: #1057088