Merged PR 233706: chore: Fix 500s thrown by heartbeat processing
**Issue**
The heartbeat controller returning unusually high volume of 500 response codes.

**Cause:**
Originally each heartbeat data handler was independently querying and updating the cloudenv record. We recently had an optimization to consolidate the changes and perform a single update.
In special cases such as an auto-suspend triggered by a handler, the suspend code path will update the cloudenv record in the db. But the controller will have a reference to an old cloudenv record which it initially pulled. When the controller tries to persist its copy of cloudenv record at the end, cosmos db was throwing an exception since the object was already updated by suspend code.

This PR makes sure that the controller's `CloudEnvironmentHandlerContext` always has the current copy of the cloudenv record.

Related work items: #1077988