Merged PR 233882: Redirect correctly with vscode-auth in VSO
This PR is meant to resolve issues with redirecting after authentication in the serverless web scenario.

A lot of this work was done by Oleg Solomka, particularly within this PR on the Cascade repo: https://devdiv.visualstudio.com/DevDiv/_git/Cascade/pullrequest/233859?_a=overview

We were able to test the vscode-account changes from the above PR integrated with the changes here and successfully login.