Merged PR 233970: feat: Adding basic Linux SKU with feature flag
This pull request is to add a 2 core basic Linux SKU with a feature flag.

The feature flag is configured in the configurations table in the cosmos db at the subscription level.
if your subscription where the plan belongs to has this feature turned on, then you will get to use this basic Linux SKU.

**NOTE:** VSO units per hour is still under discussion with management, will update that later once its finalized. As of now, I am using the standard Linux VSO units per hour value for basic Linux SKU.

## Enabling feature flag
We are following the existing pattern for configuring the feature flag in the configuration table as
id": "basicSku:isEnabled:8def34ce-053c-43ba-8501-37599fb7f010"

we then read this at run time to validate the visibility of the SKU.

![image.png](https://devdiv.visualstudio.com/579bc0c8-404b-48fd-a86e-9d65b001f1bf/_apis/git/repositories/b262268f-bbd7-4d01-b7dc-b03666e77189/pullRequests/233970/attachments/image.png)

Related work items: #1078248