Merged PR 234699: Front End generates heartbeat tokens for self-hosted envs
# What's the problem?

Self-hosted environments are unable to send heartbeats (and change environment status) because they lack the VM token that cloud environments have.

# What's new?

A new Front End API (`GET /environments/{envid}/heartbeattoken`) that will generate heartbeat tokens (VM token) for self-hosted environments. The `resourceId` claim in the token is set to the environment ID.

A `GET` to an environment ID with type `StaticEnvironment` and a valid AAD token returns:
```
{
    "heartbeatToken": "eyJhbGciOiJolll"
}
```

Using an environment ID with type `CloudEnvironment` or an invalid AAD token results in a 401.

# What's to come?

1) Changes to the heartbeat controller so that self-hosted heartbeats skip backend processing
2) Changes to the agent to generate the heartbeat token (and retain in memory)

Related work items: #1076748