Merged PR 234073: feat: VM resource marked available only after heartbeat is received.
feat: VM resource marked available only after heartbeat is received.

Prior to this change, when a VM resource is provisioned, it immediately was marked as available.
It did not wait for the heartbeats to be sent out by the VMAgent.
This sometimes results in VMs being allocated to environments, when it is not really ready.
It is more pronounced in windows environments because the VM needs to warm up different instances of Visual Studio, before it is ready to be assigned to an environment.

Related work items: #1059239