Merged PR 235670: fix: updating Sku list to have standard Linux Sku as top of the list.
This pull request contains code to update 'standard linux' as first in the list of Sku.
So that clients will have it has the default Sku. Clients as of now taking the first Sku in the list as default Sku.

The existing code has a comparator that looks for OS and VSO units per hour and sorts the lists.
With the feature of basic Sku, the default Sku got changed since it has lesser VSO units per hour.

The default setting is shown below,
![image.png](https://devdiv.visualstudio.com/579bc0c8-404b-48fd-a86e-9d65b001f1bf/_apis/git/repositories/b262268f-bbd7-4d01-b7dc-b03666e77189/pullRequests/235670/attachments/image%20(2).png)

The list is now updated with basic Sku addition shown below,
![image.png](https://devdiv.visualstudio.com/579bc0c8-404b-48fd-a86e-9d65b001f1bf/_apis/git/repositories/b262268f-bbd7-4d01-b7dc-b03666e77189/pullRequests/235670/attachments/image.png)

Related work items: #1083289