Merged PR 232855: chore: Stop clearing sessionPath on environment suspend/restart.
This PR supports: https://devdiv.visualstudio.com/DefaultCollection/DevDiv/_git/Cascade/pullrequest/231514

WHAT:
Linux environments are changing the paths where the clone is made from HOME/workspace to HOME/workspace/repository.

Existing clients use connection.sessionPath from the environment record to know the place to connect to.

Due to the latest changes the path is only known till the clone is evaluated and the original code clears this path in various places, needing it to be set again but the clone is not repeated.

There were several options to make this feature to work.

IMPLEMENTED
- The agent sends the full path the first time the environment is available.
- This path is taken as a default path, and no further changes will be made.
- Changes were made so that this entry is never cleared from the record.

Pros:
- Simple to implement.
- Doesn't require updates to clients or migration changes.
- Since the client is the one that ultimatly decides to what folder to connect to its easy to implement other mechanisms where the connection is made to other paths.
- This value can always be the fallbackpath or default path, for example if the user deletes other folders.

Cons:
- We would need to implement a specific action to update or change this path.
- If a bug clears this value then the environment will work but will default to HOME/workspace.

OTHER OPTIONS
There were other solutions each with its own pros and cons:

Saving a file inside the container or in storage that saves the path so that the monitor in the vm always has access to it.
- More complicated adding failure points between both agents.

Asking the agent inside the container what path to connect to.
- Requires client changes.

Moving this setting outside of connection on the environment record.
- Requires complicated migrations and changes to clients, services and cli.