Merged PR 237795: The react application is not getting initialized for GitHub and AzureDevOps
This started happening after the change !236942 which makes popup windows to not initialize the reactapp
But, this started happening for GitHub and AzureDevOps tabs as the isInPopupWindow will return true for new tabs as well. And it is hard to differentiate between popup and new tab from within javascript.

Related work items: #1088060