Merged PR 239215: Fix DI issues in the backend caused by Cross Region Continuation Framework
Fix DI issues in the backend caused by Cross Region Continuation Framework.

The Cross-region continuation framework is only used by the frontend at the moment, hence the related implementations need not be injected in the backend service. I have separated out the implementation classes, so that backend can stay unaware of the cross-region version of the continuation framework and continue to use the regular continuation framework.