Merged PR 238661: Remove userid references on billings objects
Today we reference and persist `userid`s in many places related to the billing, but they are not actually not used anywhere.
Unnecessary userid references on our db documents add GDPR/Privacy data compliance overhead.

This PR removes such userid references. The fields themselves are not removed from the model classes for now, so that we can use them to perform a onetime cleanup on our DB, to get rid of all the persisted userid refs.

Related work items: #1052241, #1066846