Merged PR 240300: fix: For deeplink to take the new users to environment creation panel from billing plan creation panel
As of now, we are not routing new users of DeepLink who have no plans created for the environment creation panel.
This pull request will fix that and take them back to the environment creation panel if they have deep Link params, if not they will be routed to environments page.

This is actually tracked as a [github issue](https://github.com/MicrosoftDocs/vsonline/issues/377)