Merged PR 240548: Adding list of recent folders to the environment record
This code adds a `RecentFolders` field to the environment record. This field stores a list of filepaths that will point to any additional workspaces the user opens in an environment. It also includes a function that can be called to update that list when a user opens a new workspace in an environment or chooses to remove one of the workspaces from the list.

Related work items: #1084302