Merged PR 239691: Load settings from CEDev appsettings.json last
It would be more intuitive for the CEDev copy of `appsettings.json` to load last so that any setting can be overridden. For example, the Getting Started documentation states to include the following in this file:

```
{
  "AppSettings": {
    "frontEnd": {
      "BackEndWebApiBaseAddress": "http://localhost:59228/",
      "UseMocksForLocalDevelopment": false,
      "DisableBackgroundTasksForLocalDevelopment": true,
      ...
      },
...
```

However, these three settings are also defined in `appsettings.local.json` in the repo, which takes precedence, so it's not currently possible to override them. This PR changes the ordering so that these settings would actually take effect.