Merged PR 241426: [portal][github]: add the VSCode quality query param switch for GitHub
Originally we decided to have the vscode quality set in the GitHub handshake info, but unfortunately, the handshake happens after the amdconfig.js file is loaded. Hence implementing the query param `dogfoodChannel=['stable' | 'insider']` that can specify the vscode quality. When GH lets us out of the iframe, we can put this info into the "partner info" handshake.