Merged PR 238285: Update on chart deploy files for signalR related services
We found that existing nodes on the Kubernetes cluster exhaust the CPU and so causing several services to be restarted mutliple times durign the day making the services unusables. This is especially true when a cold start happen where thousand of clients are connecting causing more activity than usual.
First of all we want to limit the CPU limit of each signalR service that was in auto mode. After observation i found out that the the servcie was assigned a number of 1900 Mi and  during work it reach around 1750 Mi. So instead of leaving auto we will request a minimum 1200 Mi and limit of 1700 Mi to avoid the node to reach high CPU limits. Also by asigning a request value for the CPU we will make sure AKS will assign each service a new node and not what we observe today that some services are put on the same node.
We also want to expand the number of nodes on the AKS from 3 -> 6 this way we can create 5 services on each node and leave 1 node to run the backplane service. By creating more servcies we will divide the number of connections per servcie and so alleviate the CPU usage per node.

Of course of all above is pure theory we can't confirm this 100% but by deploying and observing if the numbers match.