Merged PR 242315: Refactor IHttpClientProvider
For calls to the token service I need an HTTP client that does not forward the current user auth token. (It will use SP authentication instead.) I found that all HTTP client uses in the FE service were using `CurrentUserHttpClientProvider` or a subclass -- which in the case of `BackEndWebApiClient` is sort of wrong (because the BE does not authenticate users), though it causes no problems. So I'm refactoring to better support HTTP client providers that don't forward the auth token.

 - Rename `ICurrentUserHttpClientProvider` to `IHttpClientProvider` and move it from `UserProfile` to `Common.Contracts` assembly.
 - Move `ForwardingCorrelationIdHandler` from `UserProfile` to `Common.AspNetCore` assembly where it is sharable by different HTTP client provider implementations.
 - Extract the user-agent code from `CurrentUserHttpClientProvider` into `ProductInfoHeaderHandler`, also now in `Common.AspNetCore`.
 - Use DI in `CommonStartupBase` to inject the `ProductInfoHeaderValue` for the service.
 - Change `BackEndHttpClientProvider` to not inherit from `CurrentUserHttpClientProvider`, so it does not forward the current user's auth token.

Related work items: #1103681