Merged PR 245433: Switch token service back to main SP and re-enable it
The token service was failing to retrieve its SP secret, which caused it to fail to get the AAD token-decryption cert and caused any requests to it with SP credentials to be rejected with 401 Unauthorized. I discussed this with John Rivard: The deployment infrastructure we have now isn't set up to support different services in the same deployment using different SPs. So for now I'm switching the token service back to using the same SP and main keyvault as the other services. That means we don't get the separation of privileges (for defense-in-depth) that I was trying for, but it's no loss in functionality. We may revisit that problem later.

(I didn't encounter that problem in dev-stamp testing because in that environment the token service SP secret was manually configured instead of set by a deployment.)

Related work items: #1103681