Merged PR 247049: Attempt to fix the high CPU issue found on the pods running in the AKS cluster
This PR also have a fix to a regression found on the servcie related to the usage of the json-rpc lib.
There is some refactor work that was missing from the last PR around the need to use a 'ServiceInfo' entity.
Some comments in the code to highlight were the suspected high CPU consumption was spotted