Merged PR 247655: Add Ngrok Helper Utils for automatically setting the local hostname
When running `CloudEnvironmentsServices.sln` on our dev boxes, we have to have Ngrok running and modify the `CEDev/appsettings.json` file with the new Ngrok hostname, either through CLI or manually.

This PR introduces a new flag, `generateLocalHostNameFromNgrox` that will automatically set these values for you in the `AppSettings` where it's required. It does this by using the Ngrox `api/tunnels` endpoint to get the first hostname found running. If the flag is set and Ngrox is not running, it'll crash with an error. If you want to use the old behavior or need to set custom URIs, you can turn the flag off and set them yourself.