Merged PR 247976: Fix several issues found on the signalR/contacts service code around CPU/memory consumption
In our last PR we add several metrics to identify perf and events for critical methods being invoked on the contact presence service. Interesting observation gives two main culprits of the CPU consumption:
1) Requesting a subscription is the most used method and also the one who has more time spend calculating
2) Notification coming from the global backplane will attempt to resolve stub contacts with not much efficiency causing important increase on the CPU activity

For 1) the 99% of the cases used by our liveshare scenario is matching properties by email and so the code was optimized to index by that property the registered contacts on this servcie instance. In the previous case a full linear lookup was done in order to match the properties which results in a very expensive operation. The indexing by email is being performed on the existing mechanism to detect when properties on a contact has changed

For 2) a refactor on the existing mechanism to resolve stub contacts was done with the same idea in mind, the majority of the cases the stub contacts are created with just 1 match property which is if the email. We create a priority map to store the unresolved stub contacts by email.

Brief explanation on what is a stub contact:
- A stub contact is an entity created on the service that pretend to map a 'real' contact that could potential register later. Let's say a valid user register itself but want to track changes on a another contact by it's email -> 'foo@gmail.com'. At this time there is no contact being registered to that property and so we don't know which id to track the potential updates and how to exchange messages. The stub contact is created to temporary hold a valid reference to a 'real' contact. When a contact became online the service will attempt to resolve the stub contact with a truly registered contact, in this case by the initial matching email.

Other fixes include an optimization on the memory by avoiding the use of a Concurrent dictionary which require far more memory that a regular dictionary with a simple locking strategy and also considering the contention is minimal when dealing with multiple enpoints on a contact.

The last important fix is around preserving the email value on our global backplane providers, for that we need to prevent nulling the email when all the connections are removed. This strategy will allow to mimize the creation of stub contacts when they are being resolved by the backplane.

As part o fwork i added a new unit test to cover the memory and perf improvements:

here a comparission of the perf number before and after the changes:

                                                                                                    Before                                        After
Memory in MB for100k  registered contacts                                   460                                            431
Memory in MB for 100k empty/virtulized contacts                         516                                            277
Time in millisecs to request a subscription by email
for 200k (100k registered and 100k empty)                                     387                                             10