Merged PR 248647: Support prefered regions in when creating the cosmos client
After analyzing in more details the perf numbers of the servcie telemetry i realize all other regions apart from USE are taking much longer in resolving email when using the Cosmos provider.
The difference between USE: 11ms  and EUW: ~>250 ms spot me the problem that all regions are connecting directly to USE endpoint instead of the replication end points. This is because we do not configure any preferred regiosn when creatign the Cosmos Client.