Merged PR 249004: Disable the multi-region database support and improve the throttle support to avoid OOM exceptions
In a prvious commit we enable the multi-region option for our Cosmos provider. Altough it improve significantly our latency time for regions outside USE, it cause unexpected side effects when writing data when we update the contact info. We saw in telemetry waiting time of minutes causing the incoming data on the backplane to collapse the data buffers. Part of the problem was that our 'throttle' was only considerd on the buffer processing. We add also throttle support before we attempt to post a job in the queue. Otherwise we saw incoming data to increase to huge numbers (> 40k items). Our max data buffer capacity is 15k and so we will start throttle around 20k.

Here a snaphot of the telemetry showing the that suddenly the update contact just suddenly increase from 0.64 ms -> 165 secs casuing the backplane buffers to collapase.

ContactBackplaneService_UpdateContactAsync_count	ContactBackplaneService_UpdateContactAsync_avgTime
1995	0.64
6673	319068.91
16148	509461.89
16650	625381.78
16624	757946.15
16793	937671.51
17084	1065231.36
14152	1235232.31
1870	165923.15
1966	0.64
2975	0.65
2745	0.6
2631	0.65
2594	0.64