Merged PR 250484: Adding CodeTour to GitHub serverless workbench
This PR adds the CodeTour extension (a "serverless" bundled version of it) to the GitHub serverless workbench, so we can explore using it (along with rich nav) for providing a compelling code browsing/onboarding experience to repos.

To enable quick iteration, I'm currently adding this endpoint via a GitHub Gist (that I can deploy to in seconds :)), but we can migrate this to a blob endpoint next.