Merged PR 247648: Track whether or not a plan has had a final bill submitted.
Once a final bill has been submitted for a deleted plan, update the
VsoPlan by setting IsFinalBillSubmitted to true and provide a
timestamp for when the final bill was submitted.

Also added a DeletedDate property to VsoPlan so that we can start
tracking when a plan was deleted. We may use this for deciding
whether or not to archive a deleted plan in the future.